refactor: move filters to main
