+++ YouTube-client-app/src/app/core/components/header/header.component.html
@@ -3,11 +3,12 @@
     <div class="search-content">
       <img class="header__logo" src="logo.svg" alt="youtube logo" />
       <app-search-input></app-search-input>
-      <app-settings-button (click)="onShowFilter()"></app-settings-button>
+      <app-settings-button (click)="showFilter()"></app-settings-button>
     </div>
     <app-login-info></app-login-info>
   </div>
-  <!-- @if (isFilterShown) {
+</header>
+
+<!-- @if (isFilterShown) {
     <app-filter></app-filter>
   } -->
-</header>
