+++ YouTube-client-app/src/app/core/pipes/search-pipe.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { CardItemModel } from '../../../../youtube/main-page/cards-list/models/card-item.model';
+import { CardItemModel } from '../../youtube/main-page/cards-list/models/card-item.model';
 
 @Pipe({
   name: 'searchPipe',
@@ -11,6 +11,8 @@ export class SearchPipePipe implements PipeTransform {
     if (!searchQuery) {
       return cards;
     }
-    return cards.filter((card: CardItemModel) => card.snippet.title.toLowerCase().includes(searchQuery.toLowerCase()),);
+    return cards.filter((card: CardItemModel) =>
+      card.snippet.title.toLowerCase().includes(searchQuery.toLowerCase()),
+    );
   }
 }
